+++ app/corelib/src/main/java/com/inspirationlogical/receipt/corelib/repository/ProductCategoryRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProductCategoryRepository extends JpaRepository<ProductCategory, Long> {
+public interface ProductCategoryRepository extends JpaRepository<ProductCategory, Integer> {
 
     ProductCategory findByName(String name);
