+++ app/corelib/src/main/java/com/inspirationlogical/receipt/corelib/repository/ProductRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 
-    Product findById(long id);
+    Product findById(int id);
 
     List<Product> findAllByType(ProductType type);
